Add files via upload
This Model is made to predict the magnitude of an earthquake by taking into account the latitude, longitude and depth values. It uses the Multiple Regression Algorithm for this prediction. Regression analysis examines the influence of one or more independent variables on a dependent variable. Magnitude is the target attribute (dependent variable) and Latitude, Longitude and Depth form the non-target attributes (independent variables). The dataset consists of about 1000 records. Prediction of Magnitude of the earthquake can be very useful, since preventive measures and various precautions can be taken prior to the calamity, taking into account the magnitude. Hence the loss of life and property can be minimized